requirements.txt
web3==6.11.3
eth-abi==4.2.1
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
typing-extensions==4.8.0
setuptools==80.9.0
aiohttp==3.9.0
click==8.1.7
pyyaml==6.0.1 